% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSexDisaggregate.R
\name{runSexDisaggregate}
\alias{runSexDisaggregate}
\title{Launch Shiny App for sexDisaggregate}
\usage{
runSexDisaggregate()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for sexDisaggregate package.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

sexDisaggregate::runSexDisaggregate()

}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}

Silva, A. (2022) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.
}
