% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateCSV.R
\name{separateCSV}
\alias{separateCSV}
\title{Separating CSV Dataset Into Three Groups}
\usage{
separateCSV(filepath, fpName, sexTag, femaleTag, maleTag)
}
\arguments{
\item{filepath}{A character string of the filepath to find your dataset.}

\item{fpName}{A character string of CSV dataset's name. Recommand use
underscore instead of space.}

\item{sexTag}{A character string of sex column title in the dataset provided.}

\item{femaleTag}{A character string that represents "Female" in the dataset
(e.g., "F", "Female", "FEMALE", etc).}

\item{maleTag}{A character string that represents "Male" in the dataset
(e.g., "M", "Male", "MALE", etc).}
}
\value{
Returns three versions of cleaned datasets under /data directory:
(1) Cleaned version of original dataset (sex is "Female", "Male" or "NA")
(2) Female-only dataset
(3) Male-only dataset
}
\description{
Check each entry of the dataset, clean unformated sex entry to standardized
"Female"/"Male"/"NA", and output 3 cleanned versions of datasets into data/
directory.
}
\examples{
sexDisaggregate::separateCSV(filepath = "data/example_data.csv",
                               fpName = "my_data",
                               sexTag = "Sex",
                               femaleTag = "F",
                               maleTag = "M")

}
\references{
R Core Team (2022). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.
}
\author{
{Fangyi Li, \email{fangyi.li@mail.utoronto.ca}}
}
