% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDiff.R
\name{visualizeDiff}
\alias{visualizeDiff}
\title{Visualizing Sex vs. Target Category}
\usage{
visualizeDiff(dataList, yTag, visMethod)
}
\arguments{
\item{yTag}{A character string of target category title in the dataset
provided.}

\item{visMethod}{A method for visualization. Current options are "bar" or
"density".}

\item{filepath}{A character string of the filepath to find your dataset.}

\item{fpName}{A character string of CSV dataset's name. Recommand use
underscore instead of space.}

\item{sexTag}{A character string of sex column title in the dataset provided.}

\item{femaleTag}{A character string that represents "Female" in the dataset
(e.g., "F", "Female", "FEMALE", etc).}

\item{maleTag}{A character string that represents "Male" in the dataset
(e.g., "M", "Male", "MALE", etc).}
}
\value{
Returns a combined figure of
(A) Overall Percentage (without Unspecified Sex)
(B) Overall Percentage (with Unspecified Sex)
(C) General Overall Distribution
(D) Overall Distribution by Sex
}
\description{
Examine the sex percentage of the dataset in pie charts. Examine sex vs.
specific category in the dataset in either bar plot or density plot.
}
\examples{
\dontrun{

# Data preparation
exampleDataPath <- system.file("extdata",
                               "example_data.csv",
                               package = "sexDisaggregate")

dataList <- sexDisaggregate::separateCSV(filepath = exampleDataPath,
                               sexTag = "Sex",
                               femaleTag = "F",
                               maleTag = "M")

# Example 1: Target category is not a continuous and numeric value.
sexDisaggregate::visualizeDiff(dataList = dataList,
                               yTag = "Asthma",
                               visMethod = "bar")

# Example 2: Target category is a continuous and numeric value.
sexDisaggregate::visualizeDiff(dataList = dataList,
                               yTag = "Sodium Intake (mg)",
                               visMethod = "density")
}

}
\references{
R Core Team (2022). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.

Claus O. Wilke (2017). cowplot: Streamlined Plot Theme and Plot Annotations
for 'ggplot2'. URL 	https://github.com/wilkelab/cowplot.
}
\author{
{Fangyi Li, \email{fangyi.li@mail.utoronto.ca}}
}
